requirements.txt
opencv-python
timm
ftfy==6.0.3
tensorboard
tensorboardX==2.4.1
numpy==1.23
pycocoevalcap==1.2
pytorch_lightning
einops
datasets
rouge_score
transformers
numba